Adds Variable Labels to Axes and Legend
Previous version used only variable names to label the axes and legend
entries. This commit changes the program to use variable labels if they
exist, and variable names otherwise.

The one exception is the label for the y-axis if there are multiple
y-variables. I'll add in a for-loop sometime in the future to fix this.